+++ lib/generators/cms/widget/slider/slider_generator.rb
@@ -11,12 +11,9 @@ module Cms
             Model::ApiGenerator.new(behavior: behavior) do |model|
               model.name = obj_class_name
               model.title = 'Widget: Slider'
+              model.migration_path = "#{widget_path}/migrate"
+              model.model_path = model_path
               model.attributes = [
-                {
-                  name: sort_key_attribute_name,
-                  type: :string,
-                  title: 'Sort key',
-                },
                 {
                   name: images_attribute_name,
                   type: :linklist,
@@ -25,7 +22,7 @@ module Cms
               ]
             end
 
-            turn_model_into_widget(obj_class_name)
+            turn_model_into_widget(obj_class_name, model_path)
           rescue Cms::Generators::DuplicateResourceError
           end
         end
@@ -33,7 +30,7 @@ module Cms
         def copy_app_directory
           directory('app', force: true)
 
-          template('thumbnail.html.haml', 'app/widgets/slider_widget/thumbnail.html.haml')
+          template('thumbnail.html.haml', "#{widget_path}/thumbnail.html.haml")
         end
 
         def notice
@@ -44,16 +41,20 @@ module Cms
 
         private
 
+        def widget_path
+          'app/widgets/slider_widget'
+        end
+
+        def model_path
+          'app/models'
+        end
+
         def example?
           options[:example]
         end
 
         def obj_class_name
           'SliderWidget'
-        end
-
-        def sort_key_attribute_name
-          'sort_key'
         end
 
         def images_attribute_name