+++ lib/generators/cms/widget/text/text_generator.rb
@@ -11,16 +11,11 @@ module Cms
             Model::ApiGenerator.new(behavior: behavior) do |model|
               model.name = obj_class_name
               model.title = 'Widget: Text'
-              model.attributes = [
-                {
-                  name: sort_key_attribute_name,
-                  type: :string,
-                  title: 'Sort key',
-                },
-              ]
+              model.migration_path = "#{widget_path}/migrate"
+              model.model_path = model_path
             end
 
-            turn_model_into_widget(obj_class_name)
+            turn_model_into_widget(obj_class_name, model_path)
           rescue Cms::Generators::DuplicateResourceError
           end
         end
@@ -28,7 +23,7 @@ module Cms
         def create_widget
           directory('app')
 
-          template('thumbnail.html.haml', 'app/widgets/text_widget/thumbnail.html.haml')
+          template('thumbnail.html.haml', "#{widget_path}/thumbnail.html.haml")
         end
 
         def notice
@@ -39,12 +34,16 @@ module Cms
 
         private
 
-        def obj_class_name
-          'TextWidget'
+        def widget_path
+          'app/widgets/text_widget'
         end
 
-        def sort_key_attribute_name
-          'sort_key'
+        def model_path
+          'app/models'
+        end
+
+        def obj_class_name
+          'TextWidget'
         end
       end
     end