+++ lib/generators/cms/widget/video/video_generator.rb
@@ -46,12 +46,9 @@ module Cms
             Model::ApiGenerator.new(behavior: behavior) do |model|
               model.name = obj_class_name
               model.title = 'Widget: Video'
+              model.migration_path = "#{widget_path}/migrate"
+              model.model_path = model_path
               model.attributes = [
-                {
-                  name: sort_key_attribute_name,
-                  type: :string,
-                  title: 'Sort key',
-                },
                 {
                   name: video_link_attribute_name,
                   type: :linklist,
@@ -91,7 +88,7 @@ module Cms
         def copy_app_directory
           directory('app', force: true)
 
-          template('thumbnail.html.haml', 'app/widgets/video_widget/thumbnail.html.haml')
+          template('thumbnail.html.haml', "#{widget_path}/thumbnail.html.haml")
         end
 
         def notice
@@ -102,12 +99,16 @@ module Cms
 
         private
 
-        def obj_class_name
-          'VideoWidget'
+        def widget_path
+          'app/widgets/video_widget'
         end
 
-        def sort_key_attribute_name
-          'sort_key'
+        def model_path
+          'app/models'
+        end
+
+        def obj_class_name
+          'VideoWidget'
         end
 
         def video_link_attribute_name